requirements.txt
annotated-types==0.6.0
certifi==2023.7.22
charset-normalizer==3.3.0
click==8.1.7
colorama==0.4.6
idna==3.4
packaging==23.2
pydantic==2.4.2
pydantic_core==2.10.1
PyQt5==5.15.9
pyqt5-plugins==5.15.9.2.3
PyQt5-Qt5==5.15.2
PyQt5-sip==12.13.0
pyqt5-tools==5.15.9.3.3
pyqt6-plugins==6.4.2.2.3
PyQt6-Qt6==6.4.3
PyQt6-sip==13.6.0
pyqtspinner==2.0.0
python-dotenv==1.0.0
PyYAML==6.0.1
QDarkStyle==3.1
qt5-applications==5.15.2.2.3
qt5-tools==5.15.2.1.3
qt6-applications==6.4.3.2.3
qt6-tools==6.4.3.1.3
QtPy==2.4.0
requests==2.31.0
tqdm==4.66.1
typing_extensions==4.8.0
urllib3==2.0.7
